+++ accounts/views.py
@@ -1,7 +1,7 @@
 from rest_framework import viewsets
 
-from .models import Client, Company, User
-from .serializer import ClientSerializer, CompanySerializer, UserSerializer, ManagerSerializer, LoginSerializer
+from .models import Client, User
+from .serializer import ClientSerializer, UserSerializer
 
 from rest_framework.authentication import SessionAuthentication, BasicAuthentication
 from rest_framework.permissions import IsAuthenticated
@@ -11,20 +11,6 @@ from rest_framework.response import Response
 class ClientViewSet(viewsets.ModelViewSet):
     queryset = Client.objects.all()
     serializer_class = ClientSerializer
-    authentication_classes = [SessionAuthentication, BasicAuthentication]
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, format=None):
-        content = {
-            'user': str(request.user),
-            'auth': str(request.auth),
-        }
-        return Response(content)
-
-
-class CompanyViewSet(viewsets.ModelViewSet):
-    queryset = Company.objects.all()
-    serializer_class = CompanySerializer
     authentication_classes = [SessionAuthentication, BasicAuthentication]
     permission_classes = [IsAuthenticated]
 
@@ -50,32 +36,6 @@ class UserViewSet(viewsets.ModelViewSet):
         return Response(content)
 
 
-class ManagerViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = ManagerSerializer
-    authentication_classes = [SessionAuthentication, BasicAuthentication]
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, format=None):
-        content = {
-            'user': str(request.user),
-            'auth': str(request.auth),
-        }
-        return Response(content)
-
-    def get_queryset(self):
-        return super().get_queryset().filter(type = 'Manager')
 
 
-class LoginViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = LoginSerializer
-    authentication_classes = [SessionAuthentication, BasicAuthentication]
-    permission_classes = [IsAuthenticated]
 
-    def get(self, request, format=None):
-        content = {
-            'user': str(request.user),
-            'auth': str(request.auth),
-        }
-        return Response(content)
