+++ challenge/urls.py
@@ -3,8 +3,8 @@ from django.urls import path, include
 from django.urls.conf import re_path
 from rest_framework.authentication import BaseAuthentication
 from rest_framework.routers import DefaultRouter
-from accounts.views import ClientViewSet, CompanyViewSet, UserViewSet, ManagerViewSet, LoginViewSet
-
+from accounts.views import ClientViewSet, UserViewSet
+from company.views import CompanyViewSet
 
 from rest_framework import permissions
 from drf_yasg.views import get_schema_view
@@ -18,8 +18,6 @@ router = DefaultRouter()
 router.register(r'client', ClientViewSet, basename='client')
 router.register(r'company', CompanyViewSet, basename='company')
 router.register(r'user', UserViewSet, basename='user')
-router.register(r'manager', ManagerViewSet, basename='manager')
-router.register(r'login', LoginViewSet, basename='login')
 
 
 schema_view = get_schema_view(
@@ -37,8 +35,8 @@ schema_view = get_schema_view(
 
 
 urlpatterns = [
-    re_path(r'^v1/admin/', admin.site.urls),
+    re_path(r'admin/', admin.site.urls),
     re_path(r'^v1/api-auth/', include('rest_framework.urls', namespace='v1')),
-    re_path(r'^v1/', include(router.urls)),
-    re_path(r'^v1/swagger/', schema_view.with_ui('swagger', cache_timeout=0), name = 'schema-swagger-ui'),
+    re_path(r'^api/v1/', include(router.urls)),
+    re_path(r'swagger/', schema_view.with_ui('swagger', cache_timeout=0), name = 'schema-swagger-ui'),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
