+++ company/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from .models import Company
+
+
+@admin.register(Company)
+class CompanyAdmin(admin.ModelAdmin):
+    list_display = ['company_name', 'cnpj', 'address', 'city', 'status']
